+++ packages/merge-styles/src/extractStyleParts.ts
@@ -1,48 +1,54 @@
 import { IStyle, IStyleBaseArray } from './IStyle';
 import { Stylesheet } from './Stylesheet';
 
+export interface IExtractStylePartsResult {
+  classes: string[];
+  objects: IStyleBaseArray;
+}
+
 /**
  * Separates the classes and style objects. Any classes that are pre-registered
  * args are auto expanded into objects.
  */
-export function extractStyleParts(
-  ...args: (IStyle | IStyle[] | false | null | undefined)[]
-): { classes: string[]; objects: IStyleBaseArray } {
-  const classes: string[] = [];
-  const objects: {}[] = [];
+export function extractStyleParts(...args: (IStyle | IStyle[] | false | null | undefined)[]): IExtractStylePartsResult {
+  const result = {
+    classes: [],
+    objects: []
+  };
+
+  _processArgs(result, args);
+
+  return result;
+}
+
+function _processArgs<T>(result: IExtractStylePartsResult, args: (IStyle | IStyle[])[]): IExtractStylePartsResult {
   const stylesheet = Stylesheet.getInstance();
+  const { classes, objects } = result;
 
-  function _processArgs(argsList: (IStyle | IStyle[])[]): void {
-    for (const arg of argsList) {
-      if (arg) {
-        if (typeof arg === 'string') {
-          if (arg.indexOf(' ') >= 0) {
-            _processArgs(arg.split(' '));
+  for (const arg of args) {
+    if (arg) {
+      if (typeof arg === 'string') {
+        if (arg.indexOf(' ') >= 0) {
+          _processArgs(result, arg.split(' '));
+        } else {
+          const translatedArgs = stylesheet.argsFromClassName(arg);
+
+          if (translatedArgs) {
+            _processArgs(result, translatedArgs);
           } else {
-            const translatedArgs = stylesheet.argsFromClassName(arg);
-
-            if (translatedArgs) {
-              _processArgs(translatedArgs);
-            } else {
-              // Avoid adding the same class twice.
-              if (classes.indexOf(arg) === -1) {
-                classes.push(arg);
-              }
+            // Avoid adding the same class twice.
+            if (classes.indexOf(arg) === -1) {
+              classes.push(arg);
             }
           }
-        } else if (Array.isArray(arg)) {
-          _processArgs(arg);
-        } else if (typeof arg === 'object') {
-          objects.push(arg);
         }
+      } else if (Array.isArray(arg)) {
+        _processArgs(result, arg);
+      } else if (typeof arg === 'object') {
+        objects.push(arg);
       }
     }
   }
 
-  _processArgs(args);
-
-  return {
-    classes,
-    objects
-  };
+  return result;
 }
